refactor(spec): split validate.ts into map-errors + structure + orchestrator
Astral extract: keep public validate()/ValidateResult unchanged while
isolating TypeBox error mapping and data-free grammar rules. Adds
cross-stage error-order characterization so pipeline stage order cannot
drift silently.
